#161 Moving whole structures works and doesn't drop items.
On the other hand, tearing off a part does not break the entire
structure - this could lead to people running around with single-block
multiblocks if players tear off all the blocks except the controller and
take those apart.

Immersive Engineering seems to render its multiblocks in whole from one
master block, so moving parts of a multiblock is a lost cause. We'd need
to include part-by-part rendering for each block in each multiblock, and
that's a lot more trouble than it's worth.