Handle Project:Red logic tiles correctly(?) - #189
I haven't exhaustively tested this, but I think it puts things in the
right orientations. I'm not sure if it's an artifact of my test
structure, but it might be messing up the parts of the logic tile. Let
me know if anything fails to rotate properly (I need before and after
orientations to find the table entry that's wrong).